requirements.txt
Beaker==1.5.4
BeautifulSoup==3.2.1
CherryPy==3.2.2
Cython==0.22
Django==1.6.11
GitPython==0.3.2.RC1
GnuPGInterface==0.3.2
IPy==0.81
Magic-file-extensions==0.2
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.3
Paste==1.7.5.1
Pillow==2.6.2
PyPAM==0.5.0
PyXB==1.2.3
PyYAML==3.11
SPARQLWrapper==1.6.4
SSSDConfig==1.12.5
Tempita==0.5.1
Unidecode==0.04.16
astroid==1.2.1
async==0.6.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
blivet==0.61.15
boto==2.38.0
bunch==1.0.1
bz2file==0.98
chardet==2.2.1
coverage==3.7
cssselect==0.9.1
decorator==3.4.0
di==0.3
dropbox==2.2.0
dulwich==0.10.0
duplicity==0.6.25
ecdsa==0.11
fedpkg==1.20
fros==1.0
gensim==2.0.0
gitdb==0.5.4
glusterfs-api==3.5.6
google-ngram-downloader==4.0.0
gps==3.11
gyp==0.1
hg-git==0.7.0
html5lib==0.999
httplib2==0.9.1
humanize==0.5
iniparse==0.4
initial-setup==0.3.25
iotop==0.6
ipython==2.4.1
isodate==0.5.4
javapackages==4.1.0
joblib==0.8.4
jsonpointer==1.0
jsonschema==2.3.0
kitchen==1.2.1
langtable==0.0.29
lockfile==0.9.1
logilab-common==0.62.1
lxml==3.3.6
matplotlib==1.3.1
mercurial==3.0
munch==2.0.2
ndg-httpsclient==0.3.2
nltk==3.1
nose==1.3.7
ntplib==0.3.2
numpy==1.8.2
openshot==1.4.3
opster==4.1
osbs==0.14
packagedb-cli==2.9
paramiko==1.15.1
path.py==5.1
pexpect==3.1
policycoreutils-default-encoding==0.1
pwquality==1.2.4
py==1.4.31
pyOpenSSL==0.13.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycrypto==2.6.1
pycryptopp==0.6.0.1206569328141510525648634803928199668821045408958
pycurl==7.19.3.1
pygame==1.9.1release
pygobject==3.14.0
pygpgme==0.3
pyinotify==0.9.6
pykickstart==1.99.65
pyliblzma==0.5.3
pylint==1.3.1
pyparsing==1.5.7
pyparted==3.10.4
python-augeas==0.5.0
python-bugzilla==1.2.2
python-cloudfiles==1.7.10
python-dateutil==1.5
python-dmidecode==3.10.13
python-fedora==0.6.2
python-graph-core==1.8.2
python-ldap==2.4.17
python-meh==0.32
python-nss==0.16.0
pytz==2012d
pyxattr==0.5.3
pyxdg==0.25
pyzmq==14.3.1
rdfextras==0.4
rdflib==4.2.1
reportlab==3.1.8
requests==2.7.0
rpkg==1.36
rpm-python==4.12.0.1
rsa==3.1.1
scdate==1.10.9
scikit-learn==0.15.2
scipy==0.14.1
seobject==0.1
sepolicy==1.1
setroubleshoot==1.1
simplegeneric==0.8
simplejson==3.5.3
six==1.9.0
slip==0.6.4
slip.dbus==0.6.4
smart-open==1.5.2
smmap==0.8.1
uniconvertor==2.0
urlgrabber==3.10.1
urllib3==1.10.4
virtualenv==12.0.7
web.py==0.37
wsgiref==0.1.2
youtube-dl==2015.11.13
yum-langpacks==0.4.5
yum-metadata-parser==1.1.4
Beaker==1.5.4
BeautifulSoup==3.2.1
CherryPy==3.2.2
Cython==0.22
Django==1.6.11
GitPython==0.3.2.RC1
GnuPGInterface==0.3.2
IPy==0.81
Magic-file-extensions==0.2
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.3
Paste==1.7.5.1
Pillow==2.6.2
PyPAM==0.5.0
PyXB==1.2.3
PyYAML==3.11
SPARQLWrapper==1.6.4
SSSDConfig==1.12.5
Tempita==0.5.1
Unidecode==0.04.16
astroid==1.2.1
async==0.6.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
blivet==0.61.15
boto==2.38.0
bunch==1.0.1
bz2file==0.98
chardet==2.2.1
coverage==3.7
cssselect==0.9.1
decorator==3.4.0
di==0.3
dropbox==2.2.0
dulwich==0.10.0
duplicity==0.6.25
ecdsa==0.11
fedpkg==1.20
fros==1.0
gensim==2.0.0
gitdb==0.5.4
glusterfs-api==3.5.6
google-ngram-downloader==4.0.0
gps==3.11
gyp==0.1
hg-git==0.7.0
html5lib==0.999
httplib2==0.9.1
humanize==0.5
iniparse==0.4
initial-setup==0.3.25
iotop==0.6
ipython==2.4.1
isodate==0.5.4
javapackages==4.1.0
joblib==0.8.4
jsonpointer==1.0
jsonschema==2.3.0
kitchen==1.2.1
langtable==0.0.29
lockfile==0.9.1
logilab-common==0.62.1
lxml==3.3.6
matplotlib==1.3.1
mercurial==3.0
munch==2.0.2
ndg-httpsclient==0.3.2
nltk==3.1
nose==1.3.7
ntplib==0.3.2
numpy==1.8.2
openshot==1.4.3
opster==4.1
osbs==0.14
packagedb-cli==2.9
paramiko==1.15.1
path.py==5.1
pexpect==3.1
policycoreutils-default-encoding==0.1
pwquality==1.2.4
py==1.4.31
pyOpenSSL==0.13.1
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycrypto==2.6.1
pycryptopp==0.6.0.1206569328141510525648634803928199668821045408958
pycurl==7.19.3.1
pygame==1.9.1release
pygobject==3.14.0
pygpgme==0.3
pyinotify==0.9.6
pykickstart==1.99.65
pyliblzma==0.5.3
pylint==1.3.1
pyparsing==1.5.7
pyparted==3.10.4
python-augeas==0.5.0
python-bugzilla==1.2.2
python-cloudfiles==1.7.10
python-dateutil==1.5
python-dmidecode==3.10.13
python-fedora==0.6.2
python-graph-core==1.8.2
python-ldap==2.4.17
python-meh==0.32
python-nss==0.16.0
pytz==2012d
pyxattr==0.5.3
pyxdg==0.25
pyzmq==14.3.1
rdfextras==0.4
rdflib==4.2.1
reportlab==3.1.8
requests==2.7.0
rpkg==1.36
rpm-python==4.12.0.1
rsa==3.1.1
scdate==1.10.9
scikit-learn==0.15.2
scipy==0.14.1
seobject==0.1
sepolicy==1.1
setroubleshoot==1.1
simplegeneric==0.8
simplejson==3.5.3
six==1.9.0
slip==0.6.4
slip.dbus==0.6.4
smart-open==1.5.2
smmap==0.8.1
uniconvertor==2.0
urlgrabber==3.10.1
urllib3==1.10.4
virtualenv==12.0.7
web.py==0.37
wsgiref==0.1.2
youtube-dl==2015.11.13
yum-langpacks==0.4.5
yum-metadata-parser==1.1.4
